don't mention optimisation
making the runtime faster probably won't help, because most of the time
is spent on Nix evaluation anyway, and adding an entire C compiler
invocation would be counterproductive.

but generating C code would in fact be a lot tidier, because one could map
from attribute sets to nested structs directly and throw a bit of
boilerplate code at it.